refactor: 테이블 별 repository 분리
