+++ src/main/java/Application.java
@@ -1,4 +1,5 @@
-import repository.JdbcRepository;
+import repository.BoardRepository;
+import repository.GameRepository;
 import view.InputView;
 import view.OutputView;
 
@@ -7,10 +8,11 @@ public class Application {
     public static void main(String[] args) {
         InputView inputView = new InputView();
         OutputView outputView = new OutputView();
-        JdbcRepository jdbcRepository = new JdbcRepository();
+        BoardRepository boardRepository = new BoardRepository();
+        GameRepository gameRepository = new GameRepository();
 
         JanggiController janggiController = new JanggiController(inputView, outputView,
-                jdbcRepository);
+                boardRepository, gameRepository);
         janggiController.run();
     }
 }
